Correct StatusLine*NC highlights for transparent backgrounds
This code was incorrectly ported from the source repo. The
`StatusLineNC` and `StatusLineTermNC` highlights are different when the
`transparent_background_level` is set to 2, but this wasn't correctly
reflected in the code.

The reason behind this is that if `StatusLine` and `StatusLineNC` have
equal values, then (Neo)Vim will use carets (^) in the status line of
the current window. (NC stands for Non-Current.) This isn't relevant in
non-transparent mode as the highlights have different background values.

Fixes #9.